LayoutTests/fast/forms/select-ask-for-reset.html

<!DOCTYPE html>
<body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<link rel="help" href="https://html.spec.whatwg.org/multipage/forms.html#ask-for-a-reset">
<div id="log"></div>
<script>
test(function() {
    var select = document.createElement('select');
    document.body.appendChild(select);
    select.innerHTML = '<option>foo</option>';
    assert_equals(select.selectedIndex, 0);
    select.value = '';
    assert_equals(select.selectedIndex, -1);
    document.body.removeChild(select);
    // removeChild shouldn't change selection because it didn't change the
    // option list.
    assert_equals(select.selectedIndex, -1);
}, 'Removing a SELECT tree from another tree should not reset selection.');
test(function() {
    var form = document.createElement('form');
    document.body.appendChild(form);
    var select = document.createElement('select');
    form.appendChild(select);
    select.innerHTML = '<option disabled>Apple</option><option>Banana</option><option>Cherry</option>';
    assert_equals(select.selectedIndex, 1);
    form.reset();
    assert_equals(select.selectedIndex, 1);
}, 'Reset should select the first option element in the list of options that is not disabled to true');
</script>
</body>
